micro: split filesystem panel
