+++ frontend/src/app/agent/_components/filesystem-panel.tsx
@@ -1,257 +1,15 @@
 "use client";
 import { useCallback, useMemo, useRef, useState } from "react";
-import { Virtuoso } from "react-virtuoso";
-import {
-  ChevronRight,
-  ChevronDown,
-  Code,
-  File,
-  Folder,
-  Monitor,
-  Minus,
-  MessageSquarePlus,
-  PanelLeftOpen,
-  Plus,
-} from "lucide-react";
-import hljs from "highlight.js";
+import { Code, Monitor, Minus, MessageSquarePlus, PanelLeftOpen, Plus } from "lucide-react";
 import { useAppStore } from "@/store";
 import { useFilesystemPanelEffects } from "@/hooks/agent/use-filesystem-panel-effects";
 import { useTools } from "@/lib/agent/tools/context";
-import { AssistantMarkdown } from "./assistant-markdown";
-type FsEntry = {
-  name: string;
-  path: string;
-  rel: string;
-  kind: "file" | "directory";
-  size?: number;
-  modifiedAt?: string;
-};
-type Comment = {
-  id: string;
-  line: number;
-  body: string;
-  createdAt: string;
-};
-type Props = { cwd: string | null };
-const EXT_TO_LANG: Record<string, string> = {
-  js: "javascript",
-  jsx: "javascript",
-  ts: "typescript",
-  tsx: "typescript",
-  py: "python",
-  rb: "ruby",
-  rs: "rust",
-  go: "go",
-  java: "java",
-  c: "c",
-  cpp: "cpp",
-  h: "c",
-  hpp: "cpp",
-  cs: "csharp",
-  swift: "swift",
-  kt: "kotlin",
-  kts: "kotlin",
-  html: "html",
-  htm: "html",
-  svg: "xml",
-  xml: "xml",
-  xsl: "xml",
-  css: "css",
-  scss: "scss",
-  sass: "scss",
-  less: "less",
-  json: "json",
-  yaml: "yaml",
-  yml: "yaml",
-  toml: "ini",
-  sh: "bash",
-  bash: "bash",
-  zsh: "bash",
-  fish: "shell",
-  sql: "sql",
-  graphql: "graphql",
-  gql: "graphql",
-  md: "markdown",
-  mdx: "markdown",
-  dockerfile: "dockerfile",
-  makefile: "makefile",
-  lua: "lua",
-  r: "r",
-  dart: "dart",
-  zig: "zig",
-  vue: "html",
-  svelte: "html",
-};
-function languageForPath(path: string): string | undefined {
-  const name = path.split("/").pop() ?? "";
-  const lower = name.toLowerCase();
-  if (lower === "dockerfile" || lower.startsWith("dockerfile.")) return "dockerfile";
-  if (lower === "makefile" || lower.startsWith("makefile.")) return "makefile";
-  const ext = name.includes(".") ? name.split(".").pop()!.toLowerCase() : "";
-  return EXT_TO_LANG[ext];
-}
-function previewKindForPath(path: string): "html" | "jsx" | "md" | null {
-  if (/\.(html?|svg)$/i.test(path)) return "html";
-  if (/\.(jsx|tsx)$/i.test(path)) return "jsx";
-  if (/\.(md|mdx|markdown)$/i.test(path)) return "md";
-  return null;
-}
-// Infer a renderable kind from raw content (no file path available, e.g. the
-// canvas buffer). Markdown is the default since most freeform notes are prose.
-export function detectPreviewKind(content: string): "html" | "jsx" | "md" {
-  const trimmed = content.trimStart();
-  const hasMarkup = /<[A-Za-z]/.test(content);
-  if (
-    hasMarkup &&
-    (/(^|\n)\s*(import\s.+from|export\s+default|export\s+function)/.test(content) ||
-      /\bclassName=/.test(content))
-  ) {
-    return "jsx";
-  }
-  if (
-    /^<!doctype html/i.test(trimmed) ||
-    /^<html[\s>]/i.test(trimmed) ||
-    /^<(div|section|main|article|header|footer|nav|aside|h[1-6]|p|ul|ol|table|svg|body|head|span|button|a|img|figure|form|style)\b/i.test(
-      trimmed,
-    )
-  ) {
-    return "html";
-  }
-  return "md";
-}
-function extractJsxPreviewSource(source: string): string {
-  const withoutImports = source
-    .replace(/^\s*import\s.+?;?\s*$/gm, "")
-    .replace(/^\s*export\s+default\s+/gm, "")
-    .replace(/^\s*export\s+/gm, "");
-  const returnMatch = withoutImports.match(/return\s*\(([\s\S]*?)\)\s*;?\s*}/);
-  const arrowMatch = withoutImports.match(/=>\s*\(([\s\S]*?)\)\s*;?\s*$/m);
-  const body = (returnMatch?.[1] || arrowMatch?.[1] || withoutImports).trim();
-  return body
-    .replace(/\{\/\*[\s\S]*?\*\/\}/g, "")
-    .replace(/\sclassName=/g, " class=")
-    .replace(/\shtmlFor=/g, " for=")
-    .replace(/\{`([^`]+)`\}/g, "$1")
-    .replace(/\{"([^"]*)"\}/g, "$1")
-    .replace(/\{'([^']*)'\}/g, "$1")
-    .replace(/\{[^{}]*\}/g, "")
-    .replace(/<([A-Z][\w.]*)/g, '<div data-component="$1"')
-    .replace(/<\/[A-Z][\w.]*>/g, "</div>");
-}
-function previewDocument(content: string, kind: "html" | "jsx"): string {
-  const body = kind === "jsx" ? extractJsxPreviewSource(content) : content;
-  return `<!doctype html><html><head><meta charset="utf-8"><base target="_blank"><style>html,body{margin:0;padding:0}body{font:14px system-ui,-apple-system,BlinkMacSystemFont,"Segoe UI",sans-serif;color:#111;background:#fff}*{box-sizing:border-box}img,video,iframe{max-width:100%}pre,code{white-space:pre-wrap}</style></head><body>${body}</body></html>`;
-}
+import type { FileComment, FsEntry } from "@/lib/agent/filesystem-types";
+import { FileViewer } from "./filesystem-file-viewer";
+import { RenderedPreview, previewKindForOpenFile } from "./filesystem-preview";
+import { Breadcrumb, TreeFileList } from "./filesystem-tree";
 
-function TreeFileList({
-  entries,
-  searchQuery,
-  openFile,
-  onOpen,
-  onToggleDir,
-  depth,
-  expandedDirs,
-  dirChildren,
-  dirLoading,
-}: {
-  entries: FsEntry[];
-  searchQuery: string;
-  openFile: string | null;
-  onOpen: (entry: FsEntry) => void;
-  onToggleDir: (rel: string) => void;
-  depth: number;
-  expandedDirs: Set<string>;
-  dirChildren: Map<string, FsEntry[]>;
-  dirLoading: Set<string>;
-}) {
-  const filtered = useMemo(() => {
-    if (!searchQuery) return entries;
-    const q = searchQuery.toLowerCase();
-    return entries.filter((e) => e.name.toLowerCase().includes(q));
-  }, [entries, searchQuery]);
-  return (
-    <>
-      {" "}
-      {filtered.map((entry) => {
-        const isDir = entry.kind === "directory";
-        const isExpanded = expandedDirs.has(entry.rel);
-        const isLoading = dirLoading.has(entry.rel);
-        const indent = depth * 12;
-        const isActive = openFile === entry.rel;
-        const children = isDir && isExpanded ? dirChildren.get(entry.rel) : undefined;
-        return (
-          <div key={entry.path}>
-            {" "}
-            <div
-              className={`flex w-full items-center gap-1 py-0.5 text-left text-[length:var(--fs-sm)] hover:bg-(--surface) ${isActive ? "bg-(--surface) text-(--fg)" : "text-(--dim)"}`}
-              style={{ paddingLeft: `${8 + indent}px`, paddingRight: "8px" }}
-            >
-              {isDir ? (
-                <button
-                  type="button"
-                  onClick={(event) => {
-                    event.stopPropagation();
-                    onToggleDir(entry.rel);
-                  }}
-                  className="flex h-3 w-3 shrink-0 items-center justify-center"
-                  aria-label={isExpanded ? `Collapse ${entry.name}` : `Expand ${entry.name}`}
-                  title={isExpanded ? "Collapse" : "Expand"}
-                >
-                  {" "}
-                  {isLoading ? (
-                    <span className="text-[length:var(--fs-2xs)] text-(--dim)">…</span>
-                  ) : isExpanded ? (
-                    <ChevronDown className="h-3 w-3" />
-                  ) : (
-                    <ChevronRight className="h-3 w-3" />
-                  )}
-                </button>
-              ) : (
-                <span className="w-3 shrink-0" />
-              )}
-              <button
-                type="button"
-                onClick={() => (isDir ? onToggleDir(entry.rel) : onOpen(entry))}
-                title={entry.rel}
-                className="flex min-w-0 flex-1 items-center gap-1 text-left"
-              >
-                {" "}
-                {isDir ? (
-                  <Folder className="h-3 w-3 shrink-0 text-(--accent)" />
-                ) : (
-                  <File className="h-3 w-3 shrink-0" />
-                )}
-                <span className="flex-1 truncate">{entry.name}</span>{" "}
-                {!isDir && entry.size != null && entry.size > 0 ? (
-                  <span className="shrink-0 text-[length:var(--fs-2xs)] text-(--dim)">
-                    {entry.size < 1024
-                      ? `${entry.size}B`
-                      : entry.size < 1024 * 1024
-                        ? `${(entry.size / 1024).toFixed(0)}K`
-                        : `${(entry.size / (1024 * 1024)).toFixed(1)}M`}
-                  </span>
-                ) : null}
-              </button>{" "}
-            </div>
-            {children ? (
-              <TreeFileList
-                entries={children}
-                searchQuery={searchQuery}
-                openFile={openFile}
-                onOpen={onOpen}
-                onToggleDir={onToggleDir}
-                depth={depth + 1}
-                expandedDirs={expandedDirs}
-                dirChildren={dirChildren}
-                dirLoading={dirLoading}
-              />
-            ) : null}{" "}
-          </div>
-        );
-      })}
-    </>
-  );
-}
+type Props = { cwd: string | null };
 export function FilesystemPanel({ cwd }: Props) {
   const [relPath, setRelPath] = useState("");
   const [entries, setEntries] = useState<FsEntry[]>([]);
@@ -260,7 +18,7 @@ export function FilesystemPanel({ cwd }: Props) {
   const [fileTruncated, setFileTruncated] = useState(false);
   const [fileSize, setFileSize] = useState(0);
   const [loadingFile, setLoadingFile] = useState(false);
-  const [comments, setComments] = useState<Comment[]>([]);
+  const [comments, setComments] = useState<FileComment[]>([]);
   const [viewMode, setViewMode] = useState<"preview" | "code">("code");
   const [searchQuery, setSearchQuery] = useState("");
   const [expandedDirs, setExpandedDirs] = useState<Set<string>>(new Set());
@@ -349,7 +107,7 @@ export function FilesystemPanel({ cwd }: Props) {
     [dirChildren, fetchDirChildren],
   );
   const lines = useMemo(() => fileContent.split("\n"), [fileContent]);
-  const previewKind = useMemo(() => (openFile ? previewKindForPath(openFile) : null), [openFile]);
+  const previewKind = useMemo(() => previewKindForOpenFile(openFile), [openFile]);
   const addComment = useCallback(
     async (line: number, body: string) => {
       if (!cwd || !openFile || !body.trim()) return;
@@ -359,7 +117,7 @@ export function FilesystemPanel({ cwd }: Props) {
           headers: { "Content-Type": "application/json" },
           body: JSON.stringify({ cwd, path: openFile, line, body }),
         });
-        const payload = (await response.json()) as { comment?: Comment; error?: string };
+        const payload = (await response.json()) as { comment?: FileComment; error?: string };
         if (payload.comment) setComments((current) => [...current, payload.comment!]);
       } catch {
         // best-effort; comment store errors surface server-side
@@ -568,233 +326,5 @@ export function FilesystemPanel({ cwd }: Props) {
         )}
       </div>{" "}
     </div>
-  );
-}
-
-function Breadcrumb({ relPath, onRoot }: { relPath: string; onRoot: () => void }) {
-  const parts = relPath ? relPath.split("/").filter(Boolean) : [];
-  return (
-    <div className="flex h-7 shrink-0 items-center gap-0.5 overflow-x-auto px-2 text-[length:var(--fs-sm)] text-(--dim)">
-      <button
-        type="button"
-        onClick={onRoot}
-        className="shrink-0 rounded px-1 text-(--dim) hover:bg-(--surface) hover:text-(--fg)"
-        title="Project root"
-      >
-        {" "}
-        /
-      </button>{" "}
-      {parts.length > 0 && <ChevronRight className="h-3 w-3 shrink-0 text-(--dim)" />}
-      {parts.map((part, i) => (
-        <span key={i} className="flex shrink-0 items-center gap-0.5">
-          <span className="truncate font-mono text-[length:var(--fs-xs)] text-(--fg)">{part}</span>{" "}
-          {i < parts.length - 1 && <ChevronRight className="h-3 w-3 shrink-0 text-(--dim)" />}
-        </span>
-      ))}
-    </div>
-  );
-}
-
-export function RenderedPreview({
-  content,
-  kind,
-}: {
-  content: string;
-  kind: "html" | "jsx" | "md";
-}) {
-  if (kind === "md") {
-    return (
-      <div className="min-h-0 flex-1 overflow-y-auto bg-(--bg) px-3 py-2 text-sm leading-6 text-(--fg)">
-        <AssistantMarkdown text={content} />{" "}
-      </div>
-    );
-  }
-  return (
-    <iframe
-      title="Rendered file preview"
-      sandbox="allow-same-origin allow-popups allow-forms"
-      srcDoc={previewDocument(content, kind)}
-      className="min-h-0 flex-1 bg-white"
-    />
-  );
-}
-function FileViewer({
-  filePath,
-  lines,
-  fontSize,
-  comments,
-  onAddComment,
-  onRemoveComment,
-}: {
-  filePath: string;
-  lines: string[];
-  fontSize: number;
-  comments: Comment[];
-  onAddComment: (line: number, body: string) => void;
-  onRemoveComment: (id: string) => void;
-}) {
-  const [composingLine, setComposingLine] = useState<number | null>(null);
-  const [draft, setDraft] = useState("");
-  const highlightedLines = useMemo(() => {
-    const lang = languageForPath(filePath);
-    if (!lang) return null;
-    try {
-      const result = hljs.highlight(lines.join("\n"), { language: lang });
-      return result.value.split("\n");
-    } catch {
-      return null;
-    }
-  }, [filePath, lines]);
-  const commentsByLine = useMemo(() => {
-    const map = new Map<number, Comment[]>();
-    for (const comment of comments) {
-      const list = map.get(comment.line) ?? [];
-      list.push(comment);
-      map.set(comment.line, list);
-    }
-    return map;
-  }, [comments]);
-  const lineHeight = Math.round(fontSize * 1.5);
-  const submitDraft = useCallback(
-    (line: number) => {
-      const body = draft.trim();
-      if (body) onAddComment(line, body);
-      setDraft("");
-      setComposingLine(null);
-    },
-    [draft, onAddComment],
-  );
-  const renderLine = useCallback(
-    (index: number) => {
-      const lineNumber = index + 1;
-      const text = lines[index] ?? "";
-      const html = highlightedLines?.[index];
-      const lineComments = commentsByLine.get(lineNumber);
-      const composing = composingLine === lineNumber;
-      return (
-        <div className="group flex flex-col">
-          <div className="flex items-start gap-1 px-1 hover:bg-(--surface)">
-            <span
-              className="w-8 shrink-0 select-none text-right font-mono text-(--dim)"
-              style={{ fontSize: fontSize - 2, lineHeight: `${lineHeight}px` }}
-            >
-              {lineNumber}{" "}
-            </span>
-            {html ? (
-              <pre
-                className="min-w-0 flex-1 whitespace-pre font-mono text-(--fg)"
-                style={{ fontSize, lineHeight: `${lineHeight}px` }}
-                dangerouslySetInnerHTML={{ __html: html || "&nbsp;" }}
-              />
-            ) : (
-              <pre
-                className="min-w-0 flex-1 whitespace-pre font-mono text-(--fg)"
-                style={{ fontSize, lineHeight: `${lineHeight}px` }}
-              >
-                {text || "\u00a0"}
-              </pre>
-            )}
-            <button
-              type="button"
-              onClick={() => {
-                setComposingLine(composing ? null : lineNumber);
-                setDraft("");
-              }}
-              className="shrink-0 rounded p-0.5 text-(--dim) opacity-0 transition-opacity hover:text-(--fg) group-hover:opacity-100"
-              title="Comment on this line"
-              aria-label={`Comment on line ${lineNumber}`}
-              style={{ lineHeight: `${lineHeight}px` }}
-            >
-              <MessageSquarePlus className="h-3 w-3" />
-            </button>
-          </div>
-          {lineComments?.map((comment) => (
-            <div
-              key={comment.id}
-              className="ml-9 mr-2 my-0.5 flex items-start gap-2 rounded border border-(--border)/60 bg-(--surface)/50 px-2 py-1 text-[length:var(--fs-xs)] text-(--fg)/85"
-            >
-              <span className="min-w-0 flex-1 whitespace-pre-wrap break-words">{comment.body}</span>
-              <button
-                type="button"
-                onClick={() => onRemoveComment(comment.id)}
-                className="shrink-0 rounded p-0.5 text-(--dim) hover:text-(--err)"
-                title="Delete comment"
-                aria-label="Delete comment"
-              >
-                <Minus className="h-3 w-3" />
-              </button>
-            </div>
-          ))}
-          {composing ? (
-            <div className="ml-9 mr-2 my-1 flex flex-col gap-1">
-              <textarea
-                autoFocus
-                value={draft}
-                onChange={(event) => setDraft(event.target.value)}
-                onKeyDown={(event) => {
-                  if (event.key === "Enter" && (event.metaKey || event.ctrlKey)) {
-                    event.preventDefault();
-                    submitDraft(lineNumber);
-                  }
-                  if (event.key === "Escape") {
-                    setComposingLine(null);
-                    setDraft("");
-                  }
-                }}
-                placeholder="Add a comment… (⌘↵ to save)"
-                className="w-full resize-none rounded border border-(--border) bg-(--surface) px-2 py-1 text-[length:var(--fs-xs)] text-(--fg) outline-none placeholder:text-(--dim)"
-                rows={2}
-              />
-              <div className="flex justify-end gap-1">
-                <button
-                  type="button"
-                  onClick={() => {
-                    setComposingLine(null);
-                    setDraft("");
-                  }}
-                  className="rounded px-2 py-0.5 text-[length:var(--fs-xs)] text-(--dim) hover:text-(--fg)"
-                >
-                  Cancel
-                </button>
-                <button
-                  type="button"
-                  onClick={() => submitDraft(lineNumber)}
-                  className="rounded bg-(--accent)/80 px-2 py-0.5 text-[length:var(--fs-xs)] text-white hover:bg-(--accent)"
-                >
-                  Comment
-                </button>
-              </div>
-            </div>
-          ) : null}
-        </div>
-      );
-    },
-    [
-      lines,
-      highlightedLines,
-      fontSize,
-      lineHeight,
-      commentsByLine,
-      composingLine,
-      draft,
-      submitDraft,
-      onRemoveComment,
-    ],
-  );
-  if (lines.length < 2000) {
-    return (
-      <div className="min-h-0 flex-1 overflow-y-auto overflow-x-auto py-0.5">
-        {lines.map((_, index) => (
-          <div key={index}>{renderLine(index)}</div>
-        ))}{" "}
-      </div>
-    );
-  }
-  return (
-    <Virtuoso
-      className="min-h-0 flex-1"
-      totalCount={lines.length}
-      itemContent={(index) => renderLine(index)}
-    />
   );
 }
